Deduplicate model selector
