+++ GPTifier/src/command_short.cpp
@@ -2,9 +2,9 @@
 
 #include "api.hpp"
 #include "cli.hpp"
-#include "configs.hpp"
 #include "params.hpp"
 #include "parsers.hpp"
+#include "selectors.hpp"
 
 #include <fmt/core.h>
 #include <json.hpp>
@@ -12,24 +12,6 @@
 #include <string>
 
 using json = nlohmann::json;
-
-namespace {
-
-std::string select_chat_model()
-{
-#ifdef TESTING_ENABLED
-    static std::string low_cost_model = "gpt-3.5-turbo";
-    return low_cost_model;
-#endif
-
-    if (configs.chat.model.has_value()) {
-        return configs.chat.model.value();
-    }
-
-    throw std::runtime_error("No model provided via configuration file!");
-}
-
-} // namespace
 
 void command_short(int argc, char **argv)
 {
