+++ GPTifier/src/selectors.cpp
@@ -1,5 +1,6 @@
 #include "selectors.hpp"
 
+#include "configs.hpp"
 #include "utils.hpp"
 
 #include <filesystem>
@@ -62,6 +63,20 @@ std::optional<std::string> get_text_from_stdin()
 }
 
 } // namespace
+
+std::string select_chat_model()
+{
+#ifdef TESTING_ENABLED
+    static std::string low_cost_model = "gpt-3.5-turbo";
+    return low_cost_model;
+#endif
+
+    if (configs.chat.model.has_value()) {
+        return configs.chat.model.value();
+    }
+
+    throw std::runtime_error("Could not determine which model to use");
+}
 
 std::string load_input_text(const std::optional<std::string> &input_file)
 {
